requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
asgiref==3.11.1
attrs==25.4.0
backoff==2.2.1
bcrypt==5.0.0
beautifulsoup4==4.12.3
bs4==0.0.2
build==1.4.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.5
chroma-hnswlib==0.7.6
chromadb==0.5.23
click==8.3.1
cryptography==46.0.5
dataclasses-json==0.6.7
distro==1.9.0
durationpy==0.10
fastapi==0.135.1
filelock==3.25.1
flatbuffers==25.12.19
frozenlist==1.8.0
fsspec==2026.2.0
google-ai-generativelanguage==0.6.6
google-api-core==2.30.0
google-api-python-client==2.192.0
google-auth==2.49.0
google-auth-httplib2==0.3.0
google-generativeai==0.7.2
googleapis-common-protos==1.73.0
grpcio==1.78.0
grpcio-status==1.62.3
h11==0.16.0
hf-xet==1.3.2
httpcore==1.0.9
httplib2==0.31.2
httptools==0.7.1
httpx==0.28.1
huggingface_hub==0.36.2
idna==3.11
importlib_metadata==8.7.1
importlib_resources==6.5.2
iniconfig==2.3.0
jiter==0.13.0
jsonpatch==1.33
jsonpointer==3.0.0
kubernetes==35.0.0
langchain==0.2.7
langchain-chroma==0.1.2
langchain-community==0.2.7
langchain-core==0.2.43
langchain-google-genai==1.0.10
langchain-openai==0.1.16
langchain-text-splitters==0.2.4
langchainhub==0.1.20
langgraph==0.1.8
langsmith==0.1.147
markdown-it-py==4.0.0
marshmallow==3.26.2
mdurl==0.1.2
mmh3==5.2.1
mpmath==1.3.0
multidict==6.7.1
mypy_extensions==1.1.0
numpy==1.26.4
oauthlib==3.3.1
onnxruntime==1.24.3
openai==1.109.1
opentelemetry-api==1.40.0
opentelemetry-exporter-otlp-proto-common==1.40.0
opentelemetry-exporter-otlp-proto-grpc==1.40.0
opentelemetry-instrumentation==0.61b0
opentelemetry-instrumentation-asgi==0.61b0
opentelemetry-instrumentation-fastapi==0.61b0
opentelemetry-proto==1.40.0
opentelemetry-sdk==1.40.0
opentelemetry-semantic-conventions==0.61b0
opentelemetry-util-http==0.61b0
orjson==3.11.7
overrides==7.7.0
packaging==24.2
pluggy==1.6.0
posthog==7.9.8
propcache==0.4.1
proto-plus==1.27.1
protobuf==4.25.8
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pyparsing==3.3.2
pypdf==6.8.0
PyPika==0.51.1
pyproject_hooks==1.2.0
pytest==8.2.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.3
regex==2026.2.28
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.3.3
rsa==4.9.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.8.3
SQLAlchemy==2.0.48
starlette==0.52.1
sympy==1.14.0
tavily-python==0.3.4
tenacity==8.5.0
tiktoken==0.12.0
tokenizers==0.20.3
tqdm==4.67.3
typer==0.24.1
types-requests==2.32.4.20260107
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.6.3
uvicorn==0.41.0
uvloop==0.22.1
watchfiles==1.1.1
websocket-client==1.9.0
websockets==16.0
wrapt==1.17.3
yarl==1.23.0
zipp==3.23.0
